+++ Backend/src/main/java/com/github/grngoo/AutoAuctions/Controllers/UsersController.java
@@ -2,6 +2,7 @@ package com.github.grngoo.AutoAuctions.Controllers;
 
 import com.github.grngoo.AutoAuctions.DTOs.UsersDto;
 import com.github.grngoo.AutoAuctions.Models.Users;
+import com.github.grngoo.AutoAuctions.Security.JwtTokenBlacklist;
 import com.github.grngoo.AutoAuctions.Security.JwtUtility;
 import com.github.grngoo.AutoAuctions.Services.UsersService;
 import jakarta.validation.Valid;
@@ -9,8 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
 
 /**
  * Endpoints for user functionality.
@@ -26,6 +25,9 @@ public class UsersController {
 
     @Autowired
     private JwtUtility jwtUtility;
+
+    @Autowired
+    private JwtTokenBlacklist jwtTokenBlackList;
 
     /**
      * Public endpoint for signing in.
@@ -46,6 +48,25 @@ public class UsersController {
     }
 
     /**
+     * Endpoint for signing out a user.
+     * This invalidates a user's token via blacklist.
+     * Requiring users to sign back in to access protected routes.
+     *
+     * @param authorizationHeader contains the string for JWT token.
+     * @return Status of request if logout was successful or not.
+     */
+    @PostMapping("/logout")
+    public ResponseEntity<String> logoutUser(@RequestHeader("Authorization") String authorizationHeader) {
+        try {
+            String token = authorizationHeader.substring(7);
+            jwtTokenBlackList.invalidateToken(token);
+            return ResponseEntity.ok("Successfully signed out");
+        } catch (Exception e) {
+            return ResponseEntity.status(409).body("Unable to logout, invalid user token provided");
+        }
+    }
+
+    /**
      * Public endpoint for creating a new account.
      * @param usersDto Data Transmissible Object to contain request body.
      * @return Response indicating status of request.
@@ -61,14 +82,28 @@ public class UsersController {
 
     /**
      * Endpoint for deleting an account.
+     * Requires a JWT token to authenticate user.
+     * Ensures authorised users can only delete their own accounts.
+     * Also invalidate token so deleted user cannot access protected routes afterward.
+     *
      * @param usersDto Data Transmissible Object to contain request body.
+     * @param authorizationHeader contains the string for JWT token.
      * @return Response indicating status of request.
      */
     @PostMapping("/delete")
-    public ResponseEntity<String> deleteUser(@Valid @RequestBody UsersDto usersDto) {
+    public ResponseEntity<String> deleteUser(
+        @Valid @RequestBody UsersDto usersDto,
+        @RequestHeader("Authorization") String authorizationHeader
+    ) {
         try {
-            usersService.deleteUser(usersDto);
-            return ResponseEntity.ok(usersDto.getUsername() + "'s account successfully deleted");
+            String token = authorizationHeader.substring(7);
+            if (jwtUtility.validateToken(token, usersDto.getUsername())) {
+                usersService.deleteUser(usersDto);
+                jwtTokenBlackList.invalidateToken(token);
+                return ResponseEntity.ok("Account successfully deleted");
+            } else {
+                return ResponseEntity.status(401).body("Invalid credentials provided");
+            }
         } catch (Exception e) {
             return ResponseEntity.status(409).body("Unable to delete account");
         }
